genpro25/main.py
#!/usr/bin/env python3
"""
Genpro25 - Radar Data Processing and Product Generation Service

This service handles:
- BUFR file processing
- NetCDF product generation
- Radar image creation
"""

import os
import time
import logging
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


def get_db_config():
    """Get database configuration from environment variables."""
    return {
        'host': os.getenv('POSTGRES_HOST', 'db'),
        'port': os.getenv('POSTGRES_PORT', '5432'),
        'database': os.getenv('POSTGRES_DB', 'radardb'),
        'user': os.getenv('POSTGRES_USER', 'radar'),
        'password': os.getenv('POSTGRES_PASSWORD', 'radarpass'),
    }


def main():
    """Main entry point for the genpro25 service."""
    logger.info("Starting Genpro25 - Radar Data Processing Service")
    
    db_config = get_db_config()
    logger.info(f"Database host: {db_config['host']}")
    
    products_dir = os.getenv('PRODUCTS_DIR', '/data/products')
    logger.info(f"Products directory: {products_dir}")
    
    # Ensure products directory exists
    os.makedirs(products_dir, exist_ok=True)
    
    # Main processing loop
    while True:
        logger.info("Genpro25 service running... waiting for data to process")
        # TODO: Implement actual radar data processing using radarlib
        # - Monitor for new BUFR files
        # - Process BUFR to NetCDF
        # - Generate radar images
        # - Update database with product metadata
        time.sleep(60)  # Check every minute


if __name__ == "__main__":
    main()
