genpro25/requirements.txt
# Radarlib package from GitHub repository
git+https://github.com/jgmarti84/radarlib.git@main#egg=radarlib

# Database connectivity
psycopg2-binary>=2.9.9

# Configuration management
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
netCDF4>=1.6.0
